Rename to getFamilyCounts
